deploy/procedures/requirements.txt
opencv-python~=4.2.0.32
opencv-contrib-python~=4.2.0.32
torch~=1.5.1
img2vec-pytorch~=0.2.5
Pillow~=7.2.0